@@ -1,37 +1,42 @@
 {
   "name": "neurocracy-promo",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "license": "GPL-2.0",
   "repository": "github:neurocracy/promo",
-  "packageManager": "yarn@3.7.0",
+  "packageManager": "yarn@4.5.0",
+  "engines": {
+    "node": ">=20 <21",
+    "yarn": ">=4.5.0"
+  },
   "scripts": {
-    "build": "webpack --progress --config webpack.config.js",
+    "build": "webpack --progress --config webpack.config.mjs",
     "build:production": "cross-env NODE_ENV=production yarn build",
-    "build:vendor": "webpack --progress --config webpack.vendor.js",
-    "postinstall": "yarn build:vendor"
+    "build:vendor": "yarn run vendorize",
+    "postinstall": "yarn run build:vendor"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.2",
-    "@babel/preset-env": "^7.23.2",
-    "@symfony/webpack-encore": "^4.4.0",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
+    "@consensus.enterprises/pnp-vendorize": "^1.0.0",
+    "@neurocracy/postcss-easing-gradients": "^3.0.2",
+    "@symfony/webpack-encore": "^4.7.0",
     "autoprefixer": "^10.4.4",
     "copy-webpack-plugin": "^11.0.0",
-    "css-loader": "^6.8.1",
+    "css-loader": "^6.11.0",
     "favicons": "^7.1.4",
     "favicons-webpack-plugin": "^6.0.0",
-    "file-loader": "^6.0.0",
-    "glob": "^8.0.3",
-    "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.4.31",
-    "postcss-easing-gradients": "github:Ambient-Impact/postcss-easing-gradients",
+    "file-loader": "^6.2.0",
+    "glob": "^11.0.0",
+    "mini-css-extract-plugin": "^2.9.2",
+    "postcss": "^8.4.49",
     "postcss-loader": "^7.3.3",
-    "sass": "^1.43.3",
-    "sass-loader": "^13.1.0",
+    "sass": "^1.43.4 <1.79.0",
+    "sass-loader": "^13.3.3",
     "sassy-cast": "github:Ambient-Impact/SassyCast#adbbd499bb36eb0c7015ae6f23c00e3a78dc54b3",
     "sharp": "^0.32.6",
-    "svg-spritemap-webpack-plugin": "^4.5.0",
-    "webpack": "^5.88.1",
-    "webpack-cli": "^4.10.0",
+    "svg-spritemap-webpack-plugin": "^4.7.0",
+    "webpack": "^5.97.1",
+    "webpack-cli": "^6.0.1",
     "webpack-remove-empty-scripts": "^1.0.1"
   },
   "dependencies": {
@@ -40,5 +45,14 @@
     "cross-env": "^7.0.3",
     "headroom.js": "^0.12.0",
     "photoswipe": "^4.1.3"
+  },
+  "vendorize": {
+    "dirName": "web/vendor",
+    "packages": [
+      "@fontsource/exo-2",
+      "ally.js",
+      "headroom.js",
+      "photoswipe"
+    ]
   }
 }
